Create csv_data_writer.py
Python script to download stock price and volume data for a list of
ticker symbols from Yahoo Finance into a PANDAS DataFrame and write the
timestamped price and volume data to a csv file. This is a very nice way to
quickly create csv data files from a list of ticker symbols.