Create Calculate_Drawdown_Max_Drawdown
Python script to calculate the current and maximum drawdown of an asset or trading strategy. This script 
employs normalized returns to establish peak equity, current drawdown and maximum drawdown. This script is a very nice
way to calculate the current drawdown and the maximum drawdown of a PANDAS DataFrame of normalized returns. 
By employing a count that that tracks the current drawdown and maximum drawdown starting with the second data points
in the index, you can look at and compare subsequent index data points to the first index data point.

Please Note: This script will NOT run as written as a free standing script. You must specify the required library 
imports, download data and calculate the normalized returns to run this script. However, you can easily incorporate 
the code into an existing script or create a function from this script.